Cleanup and upated README
