+++ generate/run.py
@@ -1,19 +1,25 @@
 from binder.core import Generator
 
-main = Generator()
+main = Generator('../include/opencascade')
+
+# For debugging and dev
+main.bind_enums = True
+main.bind_functions = True
+main.bind_classes = True
+main.bind_typedefs = True
+main.bind_class_templates = True
 
 main.process_config('config.txt')
 
+print('Generate common header file...')
+main.generate_common_header('C:/Users/Trevor/Work/Products/pyOCCT/inc')
+
 print('Parsing headers...')
-# main.parse('all_includes.h')
-main.parse('test_all_includes.h')
-# main.dump_diagnostics()
+main.parse('all_includes.h')
+main.dump_diagnostics()
 
-print('Building modules...')
-main.build_modules()
-
-print('Grouping binders...')
-main.group_binders()
+print('Traversing headers...')
+main.traverse()
 
 print('Sorting binders...')
 main.sort_binders()
@@ -24,47 +30,13 @@ main.build_includes()
 print('Building imports...')
 main.build_imports()
 
-print('Building aliases...')
-main.build_aliases()
-
 print('Checking circular imports...')
-# main.build_include_guards()
 main.check_circular()
 
 print('Binding templates...')
-main.bind_templates(r'C:\Users\Trevor\Work\Products\pyOCCT\src\include')
+main.bind_templates('C:/Users/Trevor/Work/Products/pyOCCT/src/include')
+# main.bind_templates('include')
 
-print('Binding types...')
-main.bind(r'C:\Users\Trevor\Work\Products\pyOCCT\src\modules')
-
-# Bind a specific module
-# mod = main.get_module('Standard')
-# mod.bind(r'C:\Users\Trevor\Work\Products\pyOCCT\src')
-
-"""
-BRepFill
-BVH
-BinMXCAFDoc
-Cocoa
-FSD
-Font
-IntPatch
-IntSurf
-LDOM
-LDOMBasicString
-MeshVS
-PCDM
-PrsMgr
-ShapeFix
-ShapePersistent
-TDF
-TDocStd
-TNaming
-TObj
-TopOpeBRep
-TopOpeBRepBuild
-TopOpeBRepDS
-V3d
-VrmlData
-XmlMCAFDoc
-"""
+print('Binding...')
+main.bind('C:/Users/Trevor/Work/Products/pyOCCT/src/modules')
+# main.bind('modules')
